etl/requirement.txt
pandas
dask
mysql-connector-python
python-dotenv
